requirements.txt
streamlit==1.25.0
pandas==1.5.3
plotly==5.15.0
fpdf2==2.7.8
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==20.1.0
pyjwt==2.8.0
cryptography==41.0.7
bcrypt==4.0.1
pytesseract==0.3.10
Pillow==10.0.1
scikit-learn==1.3.0
pytest==9.0.2
opencv-python==4.12.0.88
redis==4.5.4
psycopg2-binary==2.9.7
schedule==1.2.0
psutil==5.9.5
sqlalchemy==2.0.23
xlsxwriter==3.1.9